[ISSUE#21] Add metadata (old name - report) panel
 Add metadata (old name - report) panel
     - only for classifications

URL: https://github.com/filipedu11/deeapp/issues/21
