+++ workspace/src/entities/Classification.js
@@ -1,5 +1,6 @@
 
 import { ClassificationDecode } from '../decode/ClassificationDecode';
+import { MetadaClassification } from '../panels/MetadataClassification';
 
 
 var cDec = new ClassificationDecode();
@@ -17,6 +18,7 @@ export class Classification{
         this.classificationStyle = classificationStyle;
         this.features = features;
         this.geojsonFile = geojsonFile;
+        this.metadataPanel = new MetadaClassification(this);
     }
 
     /**
@@ -112,6 +114,18 @@ export class Classification{
         return 'classification';
     }
 
+    createReport(){
+        this.metadataPanel.createReport();
+    }
+
+    getContentReport(){
+        this.metadataPanel.getContentReport();
+    }
+
+    clearReport(){
+        this.metadataPanel.clearReport();
+    }
+
     createLegend(){
         
         var legend = document.getElementById('legend');
@@ -137,67 +151,6 @@ export class Classification{
             //CLEAR PREVIOUS CONTENT
             legend.innerHTML = '';
             legend.style.display = 'none';
-        }
-    }
-
-
-    createReport(){
-        
-        var content = document.getElementById('content-metadata');
-
-        content.append(this.getContentReport());
-    }
-
-    getContentReport(){
-        var sourceC = document.getElementById('content-metadata-'+this.getId());
-
-        if (sourceC === null) {
-            sourceC = document.createElement('div');
-            sourceC.id = 'content-metadata-'+this.getId();
-        }
-
-        sourceC.innerHTML = 
-            '<div class="card border-dark mb-3">'+
-                '<div class="card-header text-center mask flex-center rgba-red-strong" data-toggle="collapse" href="#collapse' + this.getId() + '" ' +
-                    'role="button" aria-expanded="true" aria-controls="collapse">'+
-                    this.getName() +
-                '</div>' +
-                '<div class="card-body collapse" id="collapse' + this.getId() + '" style="padding: 0px">'+
-                    '<table class="table table-bordered table-dark" style="margin: 0px;">' +
-                        '<tbody>' +
-                            '<tr>' +
-                                '<th scope="row">Description</th>' +
-                                '<td>' + this.getDescription() +'</td>' +
-                            '</tr>' +
-                            '<tr>' +
-                                '<th scope="row">Author</th>'+
-                                '<td>' + this.getAuthor() + '</td>'+
-                            '</tr>'+
-                            '<tr>' +
-                                '<th scope="row">Classification Algorithm</th>'+
-                                '<td>' + this.getClassifierAlgorithm() + '</td>'+
-                            '</tr>'+
-                            '<tr>' +
-                                '<th scope="row">Collected Data Date</th>'+
-                                '<td>' + this.getCollectedDate() + '</td>'+
-                            '</tr>'+
-                            '<tr>' +
-                                '<th scope="row">Classification Date</th>'+
-                                '<td>' + this.getClassificationDate() + '</td>'+
-                            '</tr>'+
-                        '</tbody>'+
-                    '</table>' +
-                '</div>' +
-            '</div>';
-        return sourceC;
-    }
-
-    clearReport(){
-        var content = document.getElementById('content-metadata-'+this.getId());
-
-        if (content !== null) {
-            //CLEAR PREVIOUS CONTENT
-            content.innerHTML = '';
         }
     }
