+++ workspace/src/panels/Stats.js
